+++ Manifiestapp/CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.5.0
 * A MAP !!!
 * Show the events of a speaker
+* Link of event in the copy board
 
 # 2.4.1
 * Google dont like the Beeple privacy info